+++ app/src/test/java/com/sebastianvm/musicplayer/features/album/list/AlbumListStateHolderTest.kt
@@ -1,7 +1,5 @@
-package com.sebastianvm.musicplayer.ui.library.albumlist
+package com.sebastianvm.musicplayer.features.album.list
 
-import com.sebastianvm.musicplayer.features.album.list.AlbumListState
-import com.sebastianvm.musicplayer.features.album.list.AlbumListStateHolder
 import com.sebastianvm.musicplayer.features.navigation.FakeNavController
 import com.sebastianvm.musicplayer.repository.album.FakeAlbumRepositoryImpl
 import com.sebastianvm.musicplayer.repository.preferences.FakeSortPreferencesRepositoryImpl
