+++ app/src/test/java/com/sebastianvm/musicplayer/features/artist/list/ArtistListStateHolderTest.kt
@@ -1,14 +1,12 @@
-package com.sebastianvm.musicplayer.ui.library.genrelist
+package com.sebastianvm.musicplayer.features.artist.list
 
 import com.sebastianvm.musicplayer.R
-import com.sebastianvm.musicplayer.features.genre.list.GenreListState
-import com.sebastianvm.musicplayer.features.genre.list.GenreListStateHolder
-import com.sebastianvm.musicplayer.features.genre.list.GenreListUserAction
 import com.sebastianvm.musicplayer.features.navigation.FakeNavController
-import com.sebastianvm.musicplayer.repository.genre.FakeGenreRepositoryImpl
+import com.sebastianvm.musicplayer.repository.artist.FakeArtistRepositoryImpl
 import com.sebastianvm.musicplayer.repository.preferences.FakeSortPreferencesRepositoryImpl
 import com.sebastianvm.musicplayer.ui.components.lists.HeaderState
 import com.sebastianvm.musicplayer.ui.components.lists.SortButtonState
+import com.sebastianvm.musicplayer.ui.components.lists.TrailingButtonType
 import com.sebastianvm.musicplayer.ui.components.lists.toModelListItemState
 import com.sebastianvm.musicplayer.ui.util.mvvm.Data
 import com.sebastianvm.musicplayer.ui.util.mvvm.Empty
@@ -23,39 +21,41 @@ import io.kotest.core.test.TestScope
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 
-class GenreListStateHolderTest : FreeSpec({
+class ArtistListStateHolderTest : FreeSpec({
 
-    lateinit var genreRepository: FakeGenreRepositoryImpl
+    lateinit var artistRepository: FakeArtistRepositoryImpl
     lateinit var sortPreferencesRepository: FakeSortPreferencesRepositoryImpl
 
     beforeTest {
-        genreRepository = FakeProvider.genreRepository
+        artistRepository = FakeProvider.artistRepository
         sortPreferencesRepository = FakeProvider.sortPreferencesRepository
     }
 
-    fun TestScope.getSubject(): GenreListStateHolder {
-        return GenreListStateHolder(
+    fun TestScope.getSubject(): ArtistListStateHolder {
+        return ArtistListStateHolder(
             stateHolderScope = this,
-            genreRepository = genreRepository,
+            artistRepository = artistRepository,
             sortPreferencesRepository = sortPreferencesRepository,
             navController = FakeNavController()
         )
     }
 
-    "init subscribes to changes in genre list" {
+    "init subscribes to changes in artist list" {
         val subject = getSubject()
         testStateHolderState(subject) {
             awaitItem() shouldBe Loading
-            genreRepository.getGenresValue.emit(emptyList())
-            sortPreferencesRepository.getGenreListSortOrderValue.emit(MediaSortOrder.ASCENDING)
+            artistRepository.getArtistsValue.emit(emptyList())
+            sortPreferencesRepository.getArtistListSortOrderValue.emit(MediaSortOrder.ASCENDING)
             awaitItem() shouldBe Empty
 
-            val genres = FixtureProvider.genreFixtures().toList()
-            genreRepository.getGenresValue.emit(genres)
+            val artists = FixtureProvider.artistFixtures().toList()
+            artistRepository.getArtistsValue.emit(artists)
             with(awaitItem()) {
-                shouldBeInstanceOf<Data<GenreListState>>()
-                state.modelListState.items shouldBe genres.map {
-                    it.toModelListItemState()
+                shouldBeInstanceOf<Data<ArtistListState>>()
+                state.modelListState.items shouldBe artists.map {
+                    it.toModelListItemState(
+                        trailingButtonType = TrailingButtonType.More
+                    )
                 }
                 state.modelListState.headerState shouldBe HeaderState.None
             }
@@ -66,21 +66,21 @@ class GenreListStateHolderTest : FreeSpec({
         val subject = getSubject()
         testStateHolderState(subject) {
             awaitItem() shouldBe Loading
-            genreRepository.getGenresValue.emit(FixtureProvider.genreFixtures().toList())
-            sortPreferencesRepository.getGenreListSortOrderValue.emit(MediaSortOrder.ASCENDING)
+            artistRepository.getArtistsValue.emit(FixtureProvider.artistFixtures().toList())
+            sortPreferencesRepository.getArtistListSortOrderValue.emit(MediaSortOrder.ASCENDING)
             with(awaitItem()) {
-                shouldBeInstanceOf<Data<GenreListState>>()
+                shouldBeInstanceOf<Data<ArtistListState>>()
                 state.modelListState.sortButtonState shouldBe SortButtonState(
-                    text = R.string.genre_name,
+                    text = R.string.artist_name,
                     sortOrder = MediaSortOrder.ASCENDING
                 )
             }
 
-            sortPreferencesRepository.getGenreListSortOrderValue.emit(MediaSortOrder.DESCENDING)
+            sortPreferencesRepository.getArtistListSortOrderValue.emit(MediaSortOrder.DESCENDING)
             with(awaitItem()) {
-                shouldBeInstanceOf<Data<GenreListState>>()
+                shouldBeInstanceOf<Data<ArtistListState>>()
                 state.modelListState.sortButtonState shouldBe SortButtonState(
-                    text = R.string.genre_name,
+                    text = R.string.artist_name,
                     sortOrder = MediaSortOrder.DESCENDING,
                 )
             }
@@ -89,8 +89,8 @@ class GenreListStateHolderTest : FreeSpec({
 
     "sortByButtonClicked toggles sort order" {
         val subject = getSubject()
-        subject.handle(GenreListUserAction.SortByButtonClicked)
+        subject.handle(ArtistListUserAction.SortByButtonClicked)
         advanceUntilIdle()
-        sortPreferencesRepository.toggleGenreListSortOrderInvocations shouldBe listOf(listOf())
+        sortPreferencesRepository.toggleArtistListSortOrderInvocations shouldBe listOf(listOf())
     }
 })
